+++ rate_design_platform/utils/rates.py
@@ -173,3 +173,45 @@ def create_tou_rates(
         monthly_rates.append(MonthlyRateStructure(year=timestamp_dt.year, month=timestamp_dt.month, rates=month_rates))
 
     return monthly_rates
+
+
+def create_operation_schedule(
+    current_state: str, monthly_rates: list[MonthlyRateStructure], TOU_params: TOUParameters, time_step: timedelta
+) -> np.ndarray:
+    """
+    Create operational schedule based on current state
+
+    Args:
+        current_state: Current schedule state ("default" or "tou")
+        monthly_rates: List of MonthlyRateStructure, with rates for each interval in the month
+        TOU_params: TOU parameters (uses default if None)
+        time_step: Time step of the simulation
+
+    Returns:
+        Boolean array where True=operation allowed, False=restricted.
+        Length of array is sum of monthly_intervals, which is the total number of intervals in the simulation period.
+    """
+    monthly_intervals = [monthly_rate_structure.intervals for monthly_rate_structure in monthly_rates]
+
+    if current_state == "default":  # Default schedule. Operation allowed at all times.
+        return np.ones(sum(monthly_intervals), dtype=bool)
+    else:  # TOU schedule. Operation restricted during peak hours.
+        daily_peak_pattern = define_peak_hours(TOU_params, time_step)
+
+        intervals_per_day = int(HOURS_PER_DAY * SECONDS_PER_HOUR / time_step.total_seconds())
+        peak_pattern = np.array([], dtype=bool)
+        for num_intervals in monthly_intervals:
+            # Repeat pattern for the month
+            num_days = num_intervals // intervals_per_day  # Number of days in the month
+            month_pattern = np.tile(daily_peak_pattern, num_days)
+
+            # Handle remainder
+            remainder = num_intervals % intervals_per_day
+            if remainder > 0:
+                month_pattern = np.concatenate([month_pattern, daily_peak_pattern[:remainder]])
+
+            peak_pattern = np.concatenate([peak_pattern, month_pattern])
+
+        return (~peak_pattern[: sum(monthly_intervals)]).astype(
+            bool
+        )  # Operation is restricted during peak hours, hence the negation.
